# api/docs/swagger.yaml
definitions:
...
  models.ComplianceProofResponse:
    properties:
      bpan:
        type: string
      commitment:
        items:
          type: integer
        type: array
      note:
        type: string
      proof:
        items:
          type: integer
        type: array
      requirement:
        description: '"operational", "second_life", "recyclable"'
        type: string
      statement:
        description: Human-readable claim
        type: string
    type: object
...
  models.LoginPayload:
    properties:
      email:
        example: admin@example.com
        type: string
      password:
        example: MySecureP@ssw0rd!
        type: string
    type: object
  models.LoginResponseJSON:
    properties:
      message:
        example: Logged in successfully
        type: string
      role:
        example: Government Authority
        type: string
      stakeholderId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
  models.RefreshResponseJSON:
    properties:
      message:
        example: Token refreshed successfully
        type: string
    type: object
  models.RegisterStakeholderPayload:
    properties:
      aadharDocumentBase64:
        example: data:image/png;base64,iVBORw0KGgo...
        type: string
      aadharNumber:
        example: "123456789012"
        type: string
      email:
        example: admin@example.com
        type: string
      password:
        example: MySecureP@ssw0rd!
        type: string
      profileDetails:
        example: '{}'
        type: string
      role:
        example: Government Authority
        type: string
    type: object
  models.RegisterStakeholderResponseJSON:
    properties:
      stakeholderId:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      status:
        example: SUCCESS
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/batteries/{bpan}/status:
    patch:
      consumes:
      - application/json
      description: Update battery lifecycle status (service provider only)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update battery status
      tags:
      - battery
  /api/v1/batteries/{bpan}/verify/operational:
    post:
      description: Generates a ZK proof that battery meets operational standards (government/regulator
        only)
      parameters:
      - description: BPAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ComplianceProofResponse'
        "400":
          description: Bad request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Verify battery operational compliance (ZK proof)
      tags:
      - compliance
...
  /api/v1/battery:
    get:
      consumes:
      - application/json
      description: Retrieve battery information by BPAN query parameter
      parameters:
      - description: BPAN
        in: query
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get battery by query parameter
      tags:
      - battery
...
  /api/v1/manufacturers:
    get:
      description: List all registered manufacturers (admin only)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - Bearer: []
      summary: List manufacturers
      tags:
      - manufacturer
    post:
      consumes:
      - application/json
      description: Register a new manufacturer (admin only)
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Register manufacturer
      tags:
      - manufacturer
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a stakeholder and sets HttpOnly cookies
      parameters:
      - description: Login payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.LoginPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successful login
          schema:
            $ref: '#/definitions/models.LoginResponseJSON'
        "401":
          description: Invalid credentials
          schema:
            type: string
      summary: Login stakeholder
      tags:
      - auth
...
  /auth/refresh:
    post:
      description: Uses the refresh token cookie to get a new access token
      produces:
      - application/json
      responses:
        "200":
          description: Tokens refreshed
          schema:
            $ref: '#/definitions/models.RefreshResponseJSON'
        "401":
          description: Invalid or missing refresh token
          schema:
            type: string
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new stakeholder with Email and Password
      parameters:
      - description: Registration payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/models.RegisterStakeholderPayload'
      produces:
      - application/json
      responses:
        "200":
          description: Successful registration
          schema:
            $ref: '#/definitions/models.RegisterStakeholderResponseJSON'
        "400":
          description: Invalid payload
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Register a new stakeholder
      tags:
      - auth
...
  /batteries/{bpan}/verify/operational:
    post:
      description: Generates a zero-knowledge proof that battery SoH meets threshold
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify battery is operational (ZK proof)
      tags:
      - lifecycle
  /batteries/{bpan}/verify/signature:
    post:
      description: Checks cryptographic integrity of all battery data
      parameters:
      - description: Battery PAN
        in: path
        name: bpan
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Verify battery data signature
      tags:
      - lifecycle
...
swagger: "2.0"